now the operator migration seems to work properly
the main big change is that for each iteration, the Kafka consumers have to be recreated. Reason is that during migration, the new host has to set the timestamp offset explicitly. Otherwise, the consumers will receive all tuples in the topics. This only happens when the topics of the consumers of the old host vs new host differ AND the uid's of the operators do NOT differ. In other words, if we try to migrate operators that have inherent differences (different topics), but same uid's (are mapped to each other), the consumers will start from zero, unless otherwise specified. We still need to specify the offset timestamp to be right before the savepoint is created, so this was inevitable regardless.