collect metrics, new send methods and proper savepoint saving and restoration
The CollectMetrics method is still WIP, but works, and can tell exactly how many tuples it'll send the next x seconds. The new send methods are not in use, but can be either fixed or used to send tuples on and off. The savepoints are now properly saved and restored; previously, only the metadata file was sent, which works when run locally, because the metadata file points to the data files, which will be the same as long as the receiving and sending Flink instances run on the same server. However, now we zip the savepoint folder and send everything, and modified SavepointV2Serializer so that the metadata file looks for the data files in a different folder. That's the advice given in https://ci.apache.org/projects/flink/flink-docs-stable/ops/state/savepoints.html#can-i-move-the-savepoint-files-on-stable-storage when moving savepoint files on stable storage. It's not explicitly supported, but this change makes it work. It might be explicitly supported in a future Flink version.